requirements.txt
bidict
docopt
edlib
logzero
numpy
pingouin
pycountry
PyYAML
scipy
sklearn
tqdm
ufal.udpipe
unimorph
